Move doctor file
